Correciones create pagos

+++ database/migrations/2022_11_25_201747_create_ventas_table.php
@@ -26,10 +26,10 @@ return new class extends Migration
             $table->bigInteger('valorTerreno'); 
             $table->string('fechaVenta');
             $table->enum('formaVenta', ['contado', 'credito']);
-            $table->float('valorPrima')->nullable();
-            $table->float('cantidadCuotas')->nullable();
-            $table->float('valorCuotas')->nullable();
-            $table->bigInteger('valorRestantePagar')->nullable();
+            $table->double('valorPrima',20,2)->nullable();
+            $table->double('cantidadCuotas',10,2)->nullable();
+            $table->double('valorCuotas',10,2)->nullable();
+            $table->double('valorRestantePagar',20,2)->nullable();
             $table->timestamps();
         });
     }
